refactored to have less conflicting duplicate directories
