+++ app/src/main/java/com/example/rippleci/data/SchoolEvent.kt
